refactor: move createForm and initializeLoginHandler to module with appropriate name
